Interface for Stats Collection
