+++ pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkWorkloadBudgetManager.java
@@ -19,6 +19,7 @@
 package org.apache.pinot.perf;
 
 import java.util.concurrent.TimeUnit;
+import org.apache.pinot.core.accounting.DefaultWorkloadBudgetManager;
 import org.apache.pinot.core.accounting.WorkloadBudgetManager;
 import org.apache.pinot.spi.env.PinotConfiguration;
 import org.apache.pinot.spi.utils.CommonConstants;
@@ -63,7 +64,7 @@ public class BenchmarkWorkloadBudgetManager {
       config.setProperty(CommonConstants.Accounting.CONFIG_OF_WORKLOAD_ENABLE_COST_COLLECTION, true);
       config.setProperty(CommonConstants.Accounting.CONFIG_OF_WORKLOAD_ENFORCEMENT_WINDOW_MS, Long.MAX_VALUE);
 
-      _manager = new WorkloadBudgetManager(config);
+      _manager = new DefaultWorkloadBudgetManager(config);
       _manager.addOrUpdateWorkload(WORKLOAD, Long.MAX_VALUE, Long.MAX_VALUE);
     }
